Fixed Physics Raycast API using Floats not Doubles

+++ Prowl.Runtime/Physics.cs
@@ -291,7 +291,7 @@ public static class Physics
     /// <summary>
     /// Casts a ray against all colliders in the scene within a maximum distance.
     /// </summary>
-    public static bool Raycast(Vector3 origin, Vector3 direction, float maxDistance)
+    public static bool Raycast(Vector3 origin, Vector3 direction, double maxDistance)
     {
         direction = direction.normalized * maxDistance;
         var jOrigin = new JVector(origin.x, origin.y, origin.z);
@@ -305,7 +305,7 @@ public static class Physics
     /// <summary>
     /// Casts a ray against all colliders in the scene within a maximum distance and returns detailed information about the hit.
     /// </summary>
-    public static bool Raycast(Vector3 origin, Vector3 direction, float maxDistance, out RaycastHit hitInfo)
+    public static bool Raycast(Vector3 origin, Vector3 direction, double maxDistance, out RaycastHit hitInfo)
     {
         direction = direction.normalized * maxDistance;
         var jOrigin = new JVector(origin.x, origin.y, origin.z);
@@ -333,7 +333,7 @@ public static class Physics
     /// <summary>
     /// Casts a ray against all colliders in the scene with the specified layer mask.
     /// </summary>
-    public static bool Raycast(Vector3 origin, Vector3 direction, float maxDistance, LayerMask layerMask)
+    public static bool Raycast(Vector3 origin, Vector3 direction, double maxDistance, LayerMask layerMask)
     {
         direction = direction.normalized * maxDistance;
         var jOrigin = new JVector(origin.x, origin.y, origin.z);
@@ -347,7 +347,7 @@ public static class Physics
     /// <summary>
     /// Casts a ray against all colliders in the scene with the specified layer mask and returns detailed information about the hit.
     /// </summary>
-    public static bool Raycast(Vector3 origin, Vector3 direction, out RaycastHit hitInfo, float maxDistance, LayerMask layerMask)
+    public static bool Raycast(Vector3 origin, Vector3 direction, out RaycastHit hitInfo, double maxDistance, LayerMask layerMask)
     {
         direction = direction.normalized * maxDistance;
         var jOrigin = new JVector(origin.x, origin.y, origin.z);
